+++ src/services/DrugService.java
@@ -2,6 +2,11 @@ package services;
 
 import entities.Cashier;
 import entities.Drug;
+import entities.Product;
+import exceptions.NameInvalidException;
+import exceptions.ParamsProductInvalidException;
+import exceptions.ProductInvalidException;
+import exceptions.ProductNotFoundException;
 import interfaces.repositories.IDrugRepository;
 import interfaces.services.IDrugService;
 import repositories.DrugRepository;
@@ -12,23 +17,38 @@ public class DrugService implements IDrugService {
     private IDrugRepository repository = new DrugRepository();
     private ArrayList<Drug> drugs = repository.getDrugs();
     @Override
-    public void addDrug(String name, float barcode, double price, boolean is_controlled) {
-        Drug drug = new Drug(name, barcode, price, is_controlled);
-        if(drug != null){
-            repository.saveDrug(drug);
+    public void addDrug(String name, float barcode, double price, boolean is_controlled) throws ProductInvalidException, ParamsProductInvalidException {
+        if(name != null && barcode > 0 && price > 0 && is_controlled == false || is_controlled == true){
+            Drug drug = new Drug(name, barcode, price, is_controlled);
+            if(drug == null || drug.getName().isEmpty() || drug.getName().trim().isBlank() || drug.getBarcode() < 0 || drug.getPrice() < 0){
+                throw new ProductInvalidException(drug);
+            }else {
+                repository.saveDrug(drug);
+            }
+        }else{
+            throw new ParamsProductInvalidException(name,barcode,price);
         }
     }
 
     @Override
-    public void changeDrug(String name, Drug newDrug) {
+    public void changeDrug(String name, Drug newDrug) throws NameInvalidException, ProductInvalidException, ProductNotFoundException {
         Drug foundedDrug = findDrug(name);
-        if(foundedDrug != null){
+        if(name == null || name.isEmpty() || name.trim().isBlank()){
+            throw new NameInvalidException(name);
+        }else if(newDrug == null || newDrug.getName().isEmpty() || newDrug.getName().trim().isBlank() || newDrug.getBarcode() < 0 || newDrug.getPrice() < 0){
+            throw new ProductInvalidException(newDrug);
+        }else if(foundedDrug == null || foundedDrug.getName().isEmpty() || foundedDrug.getName().trim().isBlank() || foundedDrug.getBarcode() < 0 || foundedDrug.getPrice() < 0){
+            throw new ProductNotFoundException();
+        }else{
             repository.changeDrug(drugs.indexOf(foundedDrug), newDrug);
         }
     }
 
     @Override
-    public void deleteDrug(String name) {
+    public void deleteDrug(String name) throws NameInvalidException {
+        if(name == null || name.isEmpty() || name.trim().isBlank()){
+            throw new NameInvalidException(name);
+        }
         Drug foundedDrug = findDrug(name);
         if(foundedDrug != null){
             repository.removeDrug(foundedDrug);
@@ -46,9 +66,12 @@ public class DrugService implements IDrugService {
     }
 
     @Override
-    public Drug findDrug(String name) {
+    public Drug findDrug(String name) throws NameInvalidException {
+        if(name.isEmpty() || name.trim().isBlank()){
+            throw new NameInvalidException(name);
+        }
         for (Drug drug : drugs){
-            if(drug.getName() == name){
+            if(drug.getName().equals(name)){
                 return drug;
             }
         }
