+++ src/services/SaleService.java
@@ -4,6 +4,8 @@ import entities.Client;
 import entities.Employee;
 import entities.Product;
 import entities.Sale;
+import exceptions.SaleInvalidException;
+import exceptions.SaleNotFoundException;
 import interfaces.repositories.IProductRepository;
 import interfaces.repositories.ISaleRepository;
 import interfaces.services.ISaleService;
@@ -17,10 +19,15 @@ public class SaleService implements ISaleService {
     private ArrayList<Sale> sales = repository.getSales();
 
     @Override
-    public void saveSale(Employee employee, Client client, String typeOfPayament) {
-        Sale sale = new Sale(employee, client, typeOfPayament);
-        if(sale != null){
-            repository.saveSale(sale);
+    public void saveSale(Employee employee, Client client, String typeOfPayament) throws SaleInvalidException {
+        if(employee != null && client != null && typeOfPayament != null){
+            Sale sale = new Sale(employee, client, typeOfPayament);
+
+            if(sale == null || sale.getClient() == null || sale.getEmployee() == null || sale.getTypeOfPayament() == null){
+                throw new SaleInvalidException(sale);
+            }else{
+                repository.saveSale(sale);
+            }
         }
     }
 
@@ -32,7 +39,10 @@ public class SaleService implements ISaleService {
     }
 
     @Override
-    public void removeSale(Sale sale) {
+    public void removeSale(Sale sale) throws SaleNotFoundException {
+        if(sale == null){
+            throw new SaleNotFoundException();
+        }
         repository.removeSale(sale);
     }
 }
